backend/requirements/base.txt
# Django
Django==2.1.2
pytz==2018.5

# django-redis
django-redis==4.9.0
redis==2.10.6

# djangorestframework-jwt
djangorestframework-jwt==1.11.0
PyJWT==1.6.4

# djangorestframework
djangorestframework==3.9.0
six==1.11.0

# django-filter
django-filter==2.0.0

# django-rest-auth
django-rest-auth==0.9.3

# django-allauth
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
django-allauth==0.38.0
idna==2.7
oauthlib==2.1.0
python3-openid==3.1.0
requests==2.20.0
requests-oauthlib==1.0.0
urllib3==1.24

# django-extensions
django-extensions==2.1.3

# celery
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
kombu==4.2.1
vine==1.1.4

# ipython
backcall==0.1.0
decorator==4.3.0
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.6
ptyprocess==0.6.0
Pygments==2.2.0
simplegeneric==0.8.1
traitlets==4.3.2
wcwidth==0.1.7

# django-cors-headers
django-cors-headers==2.4.0
